refactor(ui): 路由基础设施并入设置「高级」tab，「高级路由」过渡 tab 移除
#165 把自动模式升为设置主入口时留了「高级路由」过渡 tab。本 PR 完成两步
走的第二步：本地路由服务（新 LocalRoutingServicePanel：开关/状态/地址
端口/日志/统计 + 首开确认）、整流器、全局出站代理三个折叠项迁入设置
「高级」tab，设置页回到 6 个 tab，路由概念彻底退居底层基础设施；
ProxyTabContent 删除，过渡 i18n（settings.tabProxy / tabAdvancedRouting）
四语清理。
